pyproject.toml
[project]
name = "nfe-parquet"
version = "0.1.0"
description = "ETL NF-e (XML/ZIP) -> Parquet mensal (AAAAMM) com checkpoint e logs JSON"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "pyyaml>=6.0",
  "pyarrow>=15.0",
  "lxml>=5.0",
  "python-dateutil>=2.9",
]

[build-system]
requires = ["hatchling>=1.20"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nfe_parquet"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]